+++ src/assets/field/src/scripts/types/Garnish.d.ts
@@ -3,6 +3,7 @@
  */
 declare const Garnish: {
   $bod: JQuery
+  Drag: GarnishComponent
   on: (target: GarnishComponent, events: string|string[], data: object|Function, handler?: Function) => void
   S_KEY: number
   uiLayerManager: {
@@ -22,4 +23,12 @@ declare interface GarnishComponent {
  */
 declare interface GarnishModal {
   show: () => void
+}
+
+declare interface GarnishDragEvent extends Event {
+  target: GarnishDragEventTarget
+}
+
+declare interface GarnishDragEventTarget extends EventTarget {
+  $draggee: JQuery<HTMLElement>
 }