+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where TinyMCE editors in Matrix, Neo and Super Table blocks became unusable after dragging the blocks
+
 ## 1.4.2 - 2023-10-26
 
 ### Changed