+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fixed a JavaScript error that occurred when editing TinyMCE fields using the 'Full' config when in preview mode
+- Fixed a bug where TinyMCE editors using the 'Full' config in Matrix, Neo and Super Table blocks became unusable after dragging the blocks
+
 ## 1.4.4 - 2023-12-10
 
 ### Fixed