+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.4.6 - 2023-12-18
 
 ### Changed
 - When (re)generating the Default and Full TinyMCE config files, the `styles` item will be used on the toolbar instead of `blocks`